+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## 2.35.1 - (April 19, 2022)
+
 * Changed
   * Updated how section header width resizing is applied. Corrects previous accidental change and incorrect wdio screenshots.
   * Updated column highlight first and last row detection for when a single row, and detecting empty and collapsed sections.